Implement standard deviation and variance aggregation functions.
Implements variance using the online algorithm described at
http://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Online_algorithm.

Standard deviation is implemented as the square root of the
variance. Aggregation functions for both population variance and
standard deviation and sample variance and standard deviation are
supplied.  Function names are postgres compatible (including a
variance and stddev aliases for var_samp and stddev_samp).