Facebook and Presto Sphinx themes
This provides two themes, the latter of which inherits from the first, and
the former of which inherits from Haiku.

Color scheme and typography is inline with contemporary Facebook look and
feel, but the overrides are quite small, so the robustness of haiku remains.

The logo is higher-resolution than shown so should work on Retina devices.