Add config option to disable interpreter fallback
The option is compiler.interpreter-enabled. When set to false, if the engine
fails to compile an expression to bytecode, the query will abort instead of
falling back to interpreted mode.

LocalQueryRunner and DistributedQueryRunner are now configured with that option
off so that we can catch certain subtle bugs in the compiler during unit testing .