Allow simple expression in VALUES
In SQL, a simple expression is a special case of a row expression with degree 1.

    'foo' = ROW ('foo')

This means that the following syntax is valid:

    VALUES 1, 2, 3, 4

and is equivalent to

    VALUES ROW (1), ROW (2), ROW (3), ROW (4)

This commit adds grammar and analyzer support for row expressions and implements
the functionality needed for VALUES to be able to work with them.