When scheduling only wait for RemoteSourceNodes
Several parts of the scheduler system would use PlanFragment.getSources
to find all nodes to schedule, and this method would look for all leaf nodes
which are not in a black list. This black list was error prone and is missing
ValuesNode.  This missing entry causes the scheduler to wait for ValuesNodes
to be scheduled which never happens since values nodes are not schduled.

The getSources is not really needed as all callers are searching for
RemoteSourceNodes, so there is simply a method for that now.